requirements/base.txt
psycopg2-binary
Django==2.1
djangorestframework==3.8.2
djangorestframework-xml==1.3.0
model_mommy
celery[redis]==4.2.0
redis==2.10.6
